fix: wrong type for read_byte size output (`uint8_t` => `long`)

+++ include/utils/file.h
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 
 
-uint8_t* read_bytes(char* path, uint8_t* size);
+uint8_t* read_bytes(char* path, long* size);
 void write_bytes(char* path, uint8_t* bytes, size_t n);
